requirements.txt
numpy==1.22.0
pandas==1.5.5
Pillow==8.4.0
matplotlib==3.5.1
seaborn==1.3.2
scikit-learn==1.1.1